Move UIKit only methods of Superview to its own file
